Fix: separation of ui logic and placeholder for troop input form
