+++ src/ui/BuildingManagementGUI.h
@@ -1,0 +1,11 @@
+#ifndef BUILDING_MANAGEMENT_GUI_H
+#define BUILDING_MANAGEMENT_GUI_H
+
+#include "../logic/PlayerResources.h"
+#include "../troops/Troop.h"
+#include <vector>
+#include <string>
+
+void displayBuildingManagementGUI(PlayerResources& playerResources, std::vector<std::string>& positions);
+
+#endif // BUILDING_MANAGEMENT_GUI_H
