+++ src/ui/GameUI.h
@@ -2,11 +2,12 @@
 #define GAME_UI_H
 
 #include "../logic/PlayerResources.h"
-#include "../troops/Troop.h"
-#include <vector>
+#include "BuildingManagementGUI.h"
+#include "TroopCreationForm.h"
+#include "TroopInputForm.h"
 #include <string>
+#include <vector>
 
-void displayTroopInputForm(std::vector<Troop> &playerTroops, std::string &selectedTarget, int &selectedLevel, std::vector<std::string> &targets, std::string &battleOutcomeMessage, PlayerResources &playerResources);
-void displayBuildingManagementGUI(PlayerResources &playerResources, std::vector<std::string> &positions); // Ensure this line is present
+void displayMainGameUI(std::vector<Troop> &playerTroops, PlayerResources &playerResources, std::vector<std::string> &targets, std::string &selectedTarget, int &selectedLevel, std::string &battleOutcomeMessage);
 
 #endif // GAME_UI_H
