+++ src/ui/TroopCreationForm.h
@@ -1,0 +1,11 @@
+#ifndef TROOP_CREATION_FORM_H
+#define TROOP_CREATION_FORM_H
+
+#include "../logic/PlayerResources.h"
+#include "../troops/Troop.h"
+#include <vector>
+#include <string>
+
+void displayTroopCreationForm(std::vector<Troop> &playerTroops, PlayerResources &playerResources);
+
+#endif // TROOP_CREATION_FORM_H
